Seems like everything is working properly
