+++ Assets/Assets/Foliage/SimpleNaturePack_Texture_01.mat.meta
