+++ Assets/Resources/FoliageColors/FoliageColor.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 90eda29f01fbb234d8f06cac849cc21a
+guid: 1bbbd15b8d78c7a4d9d793d535e55f1d
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
