+++ Assets/Resources/FoliageColors/FoliageColor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a32c17e8181c0ea4d8b4166d697c321e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
