+++ Assets/Scripts/Solar System/SolarSystemTransform.cs
@@ -36,6 +36,8 @@ public class SolarSystemTransform : MonoBehaviour
         planetsParent = gameObject;
 
         player = Universe.player;
+        string path = "FoliageColors/TintFoliageTextures";
+        Resources.Load<TintFoliageTextures>(path);
         tintFoliageTexturesHandler = GetComponent<TintFoliageTextures>();
     }
